refactor(game): move type guards
